+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [1.0.1] - 2021-03-03
 - Upgraded python-lib to v2.9.0
 
+### Fixed
+- merge: Copied the original repository into archive directory prior to processing.
+
 ### Added
 - Added -u option to allow for unrelated Git repository histories to be merged.